+++ src/components/SharedComponents/KeyStageKeypad/KeyStageKeypad.test.tsx
@@ -6,7 +6,6 @@ import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
 import keyStageKeypad from "@/browser-lib/fixtures/keyStageKeypad";
 
 const keyStages = ["EYFS", "1", "2", "3", "4"];
-const years = ["6", "7", "8", "9", "10", "11"];
 
 describe("components/Key Stage keypad", () => {
   test.each(keyStages)(
@@ -27,14 +26,5 @@ describe("components/Key Stage keypad", () => {
     const keyStageButtons = screen.getAllByRole("button");
     expect(keyStageButtons[0]).toHaveTextContent("KS1");
     expect(keyStageButtons[4]).toHaveTextContent("EYFS");
-  });
-
-  test.skip.each(years)("renders a year link with %p text", (year) => {
-    const { getByText } = renderWithTheme(
-      <KeyStageKeypad {...keyStageKeypad} />,
-    );
-    const keyStageLink = getByText(year);
-
-    expect(keyStageLink).toBeInTheDocument();
   });
 });