+++ src/components/TeacherComponents/HeaderListing/HeaderListing.tsx
@@ -51,7 +51,6 @@ export type HeaderListingProps = {
   programmeFactor: string;
   hasCurriculumDownload?: boolean;
   shareButton?: React.ReactNode;
-  copiedComponent?: React.ReactNode;
   unitDownloadFileId?: string;
   onUnitDownloadSuccess?: () => void;
   showRiskAssessmentBanner?: boolean;
@@ -59,6 +58,7 @@ export type HeaderListingProps = {
   subjectDescriptionUnitListingData?: UnitListingData;
   isUnitSaved?: boolean;
   onSave?: () => void;
+  isGeorestrictedUnit?: boolean;
 };
 
 const HeaderListing: FC<HeaderListingProps> = (props) => {
@@ -77,7 +77,6 @@ const HeaderListing: FC<HeaderListingProps> = (props) => {
     tierTitle,
     yearTitle,
     shareButton,
-    copiedComponent,
     unitDownloadFileId,
     onUnitDownloadSuccess,
     showRiskAssessmentBanner,
@@ -85,6 +84,7 @@ const HeaderListing: FC<HeaderListingProps> = (props) => {
     subjectDescriptionUnitListingData,
     isUnitSaved,
     onSave,
+    isGeorestrictedUnit,
   } = props;
 
   const isKeyStagesAvailable = keyStageSlug && keyStageTitle;
@@ -207,6 +207,7 @@ const HeaderListing: FC<HeaderListingProps> = (props) => {
                     unitFileId={unitDownloadFileId}
                     onDownloadSuccess={onUnitDownloadSuccess}
                     showNewTag={false}
+                    georestricted={Boolean(isGeorestrictedUnit)}
                   />
                 )}
                 {shareButton}
@@ -240,7 +241,6 @@ const HeaderListing: FC<HeaderListingProps> = (props) => {
                   </OakSecondaryButton>
                 )}
               </OakFlex>
-              {copiedComponent}
               <OakBox $display={["none", "block", "block"]}>
                 {bannersBlock}
               </OakBox>