+++ src/components/TeacherComponents/LessonList/LessonList.tsx
@@ -8,6 +8,10 @@ import {
   OakBox,
 } from "@oaknational/oak-components";
 
+import SignPostToAila from "../NoSearchResults/SignPostToAila";
+
+import { LessonListSeoHelper } from "./LessonListSeoHelper";
+
 import LessonListItem, {
   LessonListItemProps,
 } from "@/components/TeacherComponents/LessonListItem";
@@ -18,6 +22,7 @@ import {
 import { SpecialistLessonListItemProps } from "@/components/TeacherComponents/LessonListItem/LessonListItem";
 import { SpecialistLesson } from "@/node-lib/curriculum-api-2023/queries/specialistLessonListing/specialistLessonListing.schema";
 import { UnpublishedLessonListItem } from "@/node-lib/curriculum-api-2023/shared.schema";
+import isSlugLegacy from "@/utils/slugModifiers/isSlugLegacy";
 
 export type LessonListProps = {
   lessonCount: number;
@@ -35,9 +40,13 @@ export type LessonListProps = {
   unitTitle: string;
   onClick: (props: LessonListItemProps | SpecialistLessonListItemProps) => void;
   expiringBanner?: React.ReactNode;
+  yearTitle?: string;
+  subjectTitle?: string;
+  programmeSlug?: string;
+  examBoardSlug?: string | null;
 };
 
-const LESSONS_PER_PAGE = 5;
+const LESSONS_PER_PAGE = 20;
 
 /**
  * Contains a list of lessons
@@ -55,9 +64,23 @@ const LessonList: FC<LessonListProps> = (props) => {
     unitTitle,
     onClick,
     expiringBanner,
+    yearTitle,
+    subjectTitle,
+    subjectSlug,
+    keyStageSlug,
+    programmeSlug,
+    examBoardSlug,
   } = props;
   const { currentPage, pageSize, firstItemRef, paginationRoute } =
     paginationProps;
+
+  const showSEOAccordion =
+    programmeSlug &&
+    !isSlugLegacy(programmeSlug) &&
+    yearTitle &&
+    keyStageSlug &&
+    unitTitle &&
+    subjectTitle;
 
   return (
     <OakFlex $flexDirection="column">
@@ -71,7 +94,6 @@ const LessonList: FC<LessonListProps> = (props) => {
         </OakHeading>
         {expiringBanner}
       </OakFlex>
-
       {currentPageItems?.length ? (
         <>
           <OakUL aria-label="A list of lessons" $reset>
@@ -90,12 +112,12 @@ const LessonList: FC<LessonListProps> = (props) => {
           </OakUL>
         </>
       ) : null}
+
       {lessonCount > LESSONS_PER_PAGE ? (
         <OakBox
           $width="100%"
           $mt={["space-between-none", "auto"]}
-          $pb={["inner-padding-xl2", "inner-padding-xl4"]}
-          $pt={["inner-padding-xl4", "inner-padding-xl3"]}
+          $pt={["inner-padding-xs", "inner-padding-xl"]}
           data-testid="pagination-box"
         >
           <OakPagination
@@ -104,8 +126,30 @@ const LessonList: FC<LessonListProps> = (props) => {
             paginationHref={paginationRoute}
           />
         </OakBox>
-      ) : (
-        <OakBox $pb="inner-padding-xl2" />
+      ) : null}
+
+      <OakBox
+        $pb={["inner-padding-xl", "inner-padding-xl2"]}
+        $pt={[null, "inner-padding-m"]}
+      >
+        <SignPostToAila
+          title="Can't find what you need?"
+          text="Create a tailor-made lesson plan and resources on any topic with Aila, our free AI-powered lesson assistant. Entirely adaptable to your class and context."
+          keyStage={keyStageSlug}
+          subject={subjectSlug}
+          unitTitle={unitTitle}
+        />
+      </OakBox>
+      {showSEOAccordion && (
+        <LessonListSeoHelper
+          examBoardSlug={examBoardSlug}
+          keystageSlug={keyStageSlug}
+          programmeSlug={programmeSlug}
+          subjectTitle={subjectTitle}
+          subjectSlug={subjectSlug}
+          unitTitle={unitTitle}
+          yearTitle={yearTitle}
+        />
       )}
     </OakFlex>
   );